+++ src/test/java/org/folio/linked/data/integration/kafka/listener/handler/InventoryInstanceEventHandlerTest.java
@@ -10,7 +10,7 @@ import static org.mockito.Mockito.when;
 import java.util.Optional;
 import org.folio.linked.data.domain.dto.InventoryInstanceEvent;
 import org.folio.linked.data.domain.dto.ResourceIndexEventType;
-import org.folio.linked.data.integration.kafka.sender.search.WorkUpdateMessageSender;
+import org.folio.linked.data.integration.kafka.sender.search.work.WorkUpdateMessageSender;
 import org.folio.linked.data.model.entity.FolioMetadata;
 import org.folio.linked.data.model.entity.Resource;
 import org.folio.linked.data.repo.FolioMetadataRepository;