+++ src/test/java/org/folio/linked/data/integration/kafka/sender/search/HubReplaceMessageSenderTest.java
@@ -5,6 +5,9 @@ import static org.folio.ld.dictionary.ResourceTypeDictionary.HUB;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 
+import org.folio.linked.data.integration.kafka.sender.search.hub.HubCreateMessageSender;
+import org.folio.linked.data.integration.kafka.sender.search.hub.HubDeleteMessageSender;
+import org.folio.linked.data.integration.kafka.sender.search.hub.HubReplaceMessageSender;
 import org.folio.linked.data.model.entity.Resource;
 import org.folio.spring.testing.type.UnitTest;
 import org.junit.jupiter.api.Test;